Move app_emojis.py to correct folder
