+++ lenny/bot.py
@@ -40,7 +40,7 @@ from logger import (
 )
 from logic.config import Config
 from logic.dicecache import DiceCache
-from logic.dnd.app_emojis import get_emoji_files, init_app_emojis
+from logic.app_emojis import get_emoji_files, init_app_emojis
 from logic.favorites import FavoritesCache
 from logic.homebrew import HomebrewData
 from logic.searchcache import SearchCache
