+++ lenny/embeds/roll.py
@@ -4,7 +4,7 @@ from d20.enums import Critical
 from embeds.components import BaseModal, ModalRadioGroupComponent
 from embeds.dnd.table import DNDTableEntryView
 from embeds.embed import UserActionEmbed
-from logic.dnd.app_emojis import AppEmoji
+from logic.app_emojis import AppEmoji
 from logic.dnd.table import DNDTable, roll_table
 from logic.roll import MultiRollResult, RollResult, SingleRollResult
 from logic.voice_chat import VC, SoundType
