+++ lenny/logic/app_emojis.py
